2005-09-18  Paul Brook  <paul@codesourcery.com>
	* config/m68k/fpgnuib.c (__floatsidf): Don't rely on signed overflow.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@104383 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/m68k/fpgnulib.c
@@ -121,7 +121,7 @@ __floatsidf (long a1)
   if (a1 < 0)
     {
       sign = SIGNBIT;
-      a1 = -a1;
+      a1 = (long)-(unsigned long)a1;
       if (a1 < 0)
 	{
 	  dl.l.upper = SIGNBIT | ((32 + EXCESSD) << 20L);
